+++ projects/angular-auth-oidc-client/src/lib/config/openid-configuration.ts
@@ -29,4 +29,7 @@ export interface OpenIdConfiguration {
     storage?: any;
     customParams?: { [key: string]: string | number | boolean };
     eagerLoadAuthWellKnownEndpoints?: boolean;
+
+    // Azure B2C have implemented this incorrectly. Add support for to disable this until fixed.
+    disableRefreshIdTokenAuthTimeValidation?: boolean;
 }