+++ projects/sample-code-flow-azure-b2c/src/app/nav-menu/nav-menu.component.ts
@@ -28,6 +28,7 @@ export class NavMenuComponent implements OnInit {
     logout() {
         this.oidcSecurityService.logoff();
     }
+
     collapse() {
         this.isExpanded = false;
     }